Move create/set/get methods to `Component` class.
The methods used to be defined in the `Atmosphere` class. Move the
definition to the `Component` class to make them callable for all model
components.
